Move quick access component to discover folder
